# docs/swagger.yaml
basePath: /v1
definitions:
...
  entity.File:
    properties:
      behavior_report_id:
        type: string
      byte_entropy:
        items:
          type: integer
        type: array
      comments_count:
        type: integer
      crc32:
        type: string
      exif:
        additionalProperties:
          type: string
        type: object
      file_extension:
        type: string
      file_format:
        type: string
      first_seen:
        type: integer
      histogram:
        items:
          type: integer
        type: array
      last_scanned:
        type: integer
      magic:
        type: string
      md5:
        type: string
      ml:
        additionalProperties: true
        type: object
      multiav:
        additionalProperties: true
        type: object
      packer:
        items:
          type: string
        type: array
      pe: {}
      sha1:
        type: string
      sha256:
        type: string
      sha512:
        type: string
      size:
        type: integer
      ssdeep:
        type: string
      status:
        type: integer
      strings: {}
      submissions:
        items:
          $ref: '#/definitions/entity.Submission'
        type: array
      tags:
        additionalProperties: true
        type: object
      trid:
        items:
          type: string
        type: array
      type:
        type: string
    type: object
  entity.Submission:
    properties:
      country:
        type: string
      filename:
        type: string
      src:
        type: string
      timestamp:
        type: integer
    type: object
...
  errors.ErrorResponse:
    properties:
      details: {}
      message:
        type: string
      status:
        type: integer
    type: object
  pagination.Pages:
    properties:
      items: {}
      page:
        type: integer
      page_count:
        type: integer
      per_page:
        type: integer
      total_count:
        type: integer
    type: object
...
host: api.saferwall.com
info:
  contact:
    email: support@saferwall.com
    name: API Support
    url: https://about.saferwall.com/contact.html
  description: Interact with Saferwall Malware Analysis Platform
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://about.saferwall.com/tos
  title: Saferwall Web API
  version: "1.0"
paths:
...
  /files/:
    get:
      consumes:
      - application/json
      description: List files
      parameters:
      - description: Number of files per page
        in: query
        name: per_page
        type: integer
      - description: Specify the page number
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/pagination.Pages'
            - properties:
                items:
                  items:
                    $ref: '#/definitions/entity.File'
                  type: array
              type: object
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Retrieves a paginated list of files
      tags:
      - File
    post:
      consumes:
      - multipart/form-data
      description: Upload file for analysis.
      parameters:
      - description: binary file
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entity.File'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "413":
          description: Request Entity Too Large
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - '{}': []
        ApiKeyAuth: []
      summary: Submit a new file for scanning
      tags:
      - File
  /files/{sha256}:
    delete:
      consumes:
      - application/json
      description: Deletes a file by ID.
      parameters:
      - description: File SHA256
        in: path
        name: sha256
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/entity.File'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Deletes a file
      tags:
      - File
    get:
      consumes:
      - application/json
      description: Retrieves the content of a file report.
      parameters:
      - description: File SHA256
        in: path
        name: sha256
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.File'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Get a file report
      tags:
      - File
    head:
      description: Check weather a file exists in the database.
      parameters:
      - description: File SHA256
        in: path
        name: sha256
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.File'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Check if a file exists.
      tags:
      - File
    patch:
      consumes:
      - application/json
      description: Patch a portion of a file report.
      parameters:
      - description: File SHA256
        in: path
        name: sha256
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.File'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Update a file report (partial update)
      tags:
      - File
    put:
      consumes:
      - application/json
      description: Replace a file report with a new report
      parameters:
      - description: File SHA256
        in: path
        name: sha256
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.File'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Update a file report (full update)
      tags:
      - File
  /files/{sha256}/comments/:
    get:
      description: List of comments for a given file.
      parameters:
      - description: File SHA256
        in: path
        name: sha256
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pagination.Pages'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Returns a paginated list of file comments
      tags:
      - File
  /files/{sha256}/download/:
    get:
      description: Download a binary file. Files are in zip format and password protected.
      parameters:
      - description: File SHA256
        in: path
        name: sha256
        required: true
        type: string
      produces:
      - multipart/form-data
      responses:
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Download a file
      tags:
      - File
  /files/{sha256}/generate-presigned-url/:
    post:
      description: Generate a pre-signed URL to download samples directly from the
        object storage.
      parameters:
      - description: File SHA256
        in: path
        name: sha256
        required: true
        type: string
      produces:
      - application/json
      responses:
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Generate a pre-signed URL for downloading samples.
      tags:
      - File
  /files/{sha256}/like/:
    post:
      description: Adds a file to the like list.
      parameters:
      - description: File SHA256
        in: path
        name: sha256
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Like a file
      tags:
      - File
  /files/{sha256}/rescan/:
    post:
      description: Rescan an existing file.
      parameters:
      - description: File SHA256
        in: path
        name: sha256
        required: true
        type: string
      produces:
      - application/json
      responses:
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Rescan an existing file
      tags:
      - File
  /files/{sha256}/strings/:
    get:
      consumes:
      - application/json
      description: List strings of a file.
      parameters:
      - description: File SHA256
        in: path
        name: sha256
        required: true
        type: string
      - description: Number of strings per page
        in: query
        name: per_page
        type: integer
      - description: Specify the page number
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pagination.Pages'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Returns a paginated list of strings
      tags:
      - File
  /files/{sha256}/summary/:
    get:
      description: File metadata returned in the summary view of a file.
      parameters:
      - description: File SHA256
        in: path
        name: sha256
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pagination.Pages'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: File summary and metadata
      tags:
      - File
  /files/{sha256}/unlike/:
    post:
      description: Removes a file from the like list.
      parameters:
      - description: File SHA256
        in: path
        name: sha256
        required: true
        type: string
      produces:
      - application/json
      responses:
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Unlike a file
      tags:
      - File
...
schemes:
- https
securityDefinitions:
  Bearer:
    description: 'Enter the token with the `Bearer: ` prefix, e.g. "Bearer abcde12345".'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"